added window.link and reorganized the axios calls
