+++ src/axios-services/index.js
@@ -1,4 +1,5 @@
 import axios from 'axios';
+const BASE_URL = "/api";
 
 // this file holds your frontend network request adapters
 // think about each function as a service that provides data
@@ -27,3 +28,25 @@ export async function getAPIHealth() {
     return { healthy: false };
   }
 }
+
+export const fetchOrder = async (id) => {
+  try {
+    const { data: order } = await axios.get(`${BASE_URL}/orders/:orderId`);
+    console.log("order:", order);
+    return order;
+  } catch (error) {
+    console.error(error);
+  }
+};
+
+export const createPendingOrder = async (email, address) => {
+  try {
+    const {
+      data: { pendingOrder },
+    } = await axios.post(`${BASE_URL}/orders`, { email, address });
+    console.log("pendingOrder", pendingOrder);
+    return pendingOrder;
+  } catch (error) {
+    console.error(error);
+  }
+};
